introduce the new user service
